+++ src/features/transactions/data/transaction.ts
@@ -6,7 +6,7 @@ import { TransactionId } from './types'
 import { asTransaction } from '../mappers/transaction-mappers'
 import { getTransactionResultAtom } from './transaction-result'
 import { assetSummaryResolver } from '@/features/assets/data/asset-summary'
-import { abiMethodResolver } from '@/features/abi-methods/get-abi-method-atom'
+import { abiMethodResolver } from '@/features/abi-methods/data/abi-method'
 
 export const createTransactionsAtom = (transactionResults: TransactionResult[]) => {
   return atom((_) => {
